pyproject.toml
[project]
name = "data-agent"
version = "0.1.0"
description = "AI Data Agent for PostgreSQL, GA4 and report generation"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary]>=3.2.0",
    "psycopg-pool>=3.2.0",
    "pandas>=2.2.0",
    "google-analytics-data>=0.18.0",
    "jinja2>=3.1.0",
    "matplotlib>=3.9.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/data_agent"]

[tool.ruff]
line-length = 120
